Added google benchmark to getlibs

+++ getlibs.sh
@@ -77,7 +77,7 @@ for target in "$@" ; do
         echo "Building $LIBRARY in $BUILD from $SOURCE"
         pushd "$SOURCE_DIR"
         if [ ! -r "$LIBRARY.tar.xz" ]; then
-            #  wget "https://github.com/boostorg/boost/releases/download/{$PACK%.tar.xz}/$PACK" -cO "$PACK"
+            # wget "https://github.com/boostorg/boost/releases/download/$LIBRARY/$LIBRARY.tar.xz" -cO "$LIBRARY.tar.xz"
             wget "https://homes.cs.aau.dk/~marius/mirrors/boost/$LIBRARY.tar.xz" -cO "$LIBRARY.tar.xz"
         fi
         if [ ! -d "$SOURCE" ]; then
@@ -85,6 +85,29 @@ for target in "$@" ; do
         fi
         mkdir -p "$BUILD"
         cmake -S "$SOURCE" -B "$BUILD" $CMAKE_ARGS -DCMAKE_INSTALL_PREFIX="$PREFIX_DIR" -DBUILD_SHARED_LIBS=OFF -DBOOST_INCLUDE_LIBRARIES="headers;math" -DBOOST_ENABLE_MPI=OFF -DBOOST_ENABLE_PYTHON=OFF -DBOOST_RUNTIME_LINK=static -DBUILD_TESTING=OFF -DBOOST_USE_STATIC_LIBS=ON -DBOOST_USE_DEBUG_LIBS=ON -DBOOST_USE_RELEASE_LIBS=ON -DBOOST_USE_STATIC_RUNTIME=ON
+        cmake --build "$BUILD" --config Release
+        cmake --install "$BUILD" --config Release
+        rm -Rf "$BUILD"
+    fi
+
+    ## Google Benchmark
+    LIBRARY="benchmark-1.7.1"
+    VERSION=${LIBRARY/[^-]*-/}
+    SOURCE="$SOURCE_DIR/$LIBRARY"
+    BUILD="$PREFIX_DIR/tmp/$LIBRARY"
+    if [ -r "$PREFIX_DIR/include/benchmark/benchmark.h" ] ; then
+        echo "$LIBRARY is already installed in $PREFIX_DIR"
+    else
+        echo "Building $LIBRARY in $BUILD from $SOURCE"
+        pushd "$SOURCE_DIR"
+        if [ ! -r "$LIBRARY.tar.gz" ]; then
+            wget "https://github.com/google/benchmark/archive/refs/tags/v$VERSION.tar.gz" -cO "$LIBRARY.tar.gz"
+        fi
+        if [ ! -d "$SOURCE" ]; then
+            tar xf "$LIBRARY.tar.gz"
+        fi
+        mkdir -p "$BUILD"
+        cmake -S "$SOURCE" -B "$BUILD" $CMAKE_ARGS -DCMAKE_INSTALL_PREFIX="$PREFIX_DIR" -DBUILD_SHARED_LIBS=OFF -DBENCHMARK_ENABLE_TESTING=OFF -DBENCHMARK_ENABLE_EXCEPTIONS=ON -DBENCHMARK_ENABLE_LTO=OFF -DBENCHMARK_USE_LIBCXX=OFF -DBENCHMARK_ENABLE_WERROR=ON -DBENCHMARK_FORCE_WERROR=OFF
         cmake --build "$BUILD" --config Release
         cmake --install "$BUILD" --config Release
         rm -Rf "$BUILD"
